Documentation Reference

Stability

The entries written on this page are automatically generated and DO NOT represent the currently supported API surface. Feel free to use anything you can find here, but be aware that just because it's listed here, does not mean it's covered under semver.

Data reference

The Data module contains most everyday objects you're going to use when writing property based tests with Supposition.jl. For example, the basic generators for integers, strings, floating point values etc. are defined here. Everything listed in this section is considered supported under semver.

Supposition.Data.postype

Supposition.Data.postype

Supposition.Data.AsciiCharacters

Supposition.Data.Bind

Supposition.Data.Booleans

Supposition.Data.Characters

Supposition.Data.Floats

Supposition.Data.Integers

Supposition.Data.Just

Supposition.Data.Map

Supposition.Data.Possibility

Supposition.Data.SampledFrom

Supposition.Data.Satisfying

Supposition.Data.Text

Supposition.Data.Vectors

Functions

Supposition.Data.postype — Method

postype(::P) where P <: Possibility

Gives the type of objects this Possibility object will generate.

Supposition.Data.postype — Method

postype(::Type{P<:Possibility})

Gives the type of objects this Possibility type will generate.

Supposition.Data.AsciiCharacters — Type

AsciiCharacters() <: Possibility{Char}

A Possibility of producing arbitrary Char instances that are isascii. More efficient than filtering Characters.

julia> using Supposition

julia> ascii = Data.AsciiCharacters()

julia> example(ascii, 5)
5-element Vector{Char}:
 '8': ASCII/Unicode U+0038 (category Nd: Number, decimal digit)
 'i': ASCII/Unicode U+0069 (category Ll: Letter, lowercase)
 'R': ASCII/Unicode U+0052 (category Lu: Letter, uppercase)
 '\f': ASCII/Unicode U+000C (category Cc: Other, control)
 '>': ASCII/Unicode U+003E (category Sm: Symbol, math)

Supposition.Data.Bind — Type

Bind(source::Possibility, f)

Binds f to source, i.e., on produce(::Bind, ::TestCase) this calls produce on source, the result of which is passed to f, the output of which will be used as input to produce again.

In other words, f takes a value produced by source and gives back a Possibility that is then immediately produced from.

Equivalent to bind(f, source).

Supposition.Data.Booleans — Type

Booleans() <: Possibility{Bool}

A Possibility for sampling boolean values.

julia> using Supposition

julia> bools = Data.Booleans()

julia> example(bools, 4)
4-element Vector{Bool}:
 0
 1
 0
 1

Supposition.Data.Characters — Type

Characters(;valid::Bool = false) <: Possibility{Char}

A Possibility of producing arbitrary Char instances.

Unicode

This will produce ANY possible Char by default, not just valid unicode codepoints! To only produce valid unicode codepoints, pass valid=true as a keyword argument.

julia> using Supposition

julia> chars = Data.Characters()

julia> example(chars, 5)
5-element Vector{Char}:
 '⠺': Unicode U+283A (category So: Symbol, other)
 '𰳍': Unicode U+30CCD (category Lo: Letter, other)
 '\U6ec9c': Unicode U+6EC9C (category Cn: Other, not assigned)
 '\U1a05c5': Unicode U+1A05C5 (category In: Invalid, too high)
 '𓂫': Unicode U+130AB (category Lo: Letter, other)

Supposition.Data.Floats — Type

Floats{T <: Union{Float16,Float32,Float64}} <: Possibility{T}

A Possibility for sampling floating point values.

Inf, Nan

This possibility will generate any valid instance, including positive and negative infinities, signaling and quiet NaNs and every possible float.

julia> using Supposition

julia> floats = Data.Floats{Float16}()

julia> example(floats, 5)
5-element Vector{Float16}:
  -8.3e-6
   1.459e4
   3.277
 NaN
  -0.0001688

Supposition.Data.Integers — Type

Integers(minimum::T, maximum::T) <: Possibility{T <: Integer}
Integers{T}() <: Possibility{T <: Integer}

A Possibility representing drawing integers from [minimum, maximum]. The second constructors draws from the entirety of T.

Produced values are of type T.

julia> using Supposition

julia> is = Data.Integers{Int}()

julia> example(is, 5)
5-element Vector{Int64}:
 -5854403925308846160
  4430062772779972974
    -9995351034504801
  2894734754389242339
 -6640496903289665416

Supposition.Data.Just — Type

Just(value::T) <: Possibility{T}

A Possibility that always produces value.

Mutable Data

The source object given to this Just is not copied when produce is called. Be careful with mutable data!

julia> using Supposition

julia> three = Data.Just(3)

julia> example(three, 3)
3-element Vector{Int64}:
 3
 3
 3

Supposition.Data.Map — Type

Map(source::Possibility, f) <: Possibility

A Possibility representing mapping values from source through f.

Equivalent to calling map(f, source).

The pre-calculated return type of Map is a best effort and may be wider than necessary.

julia> using Supposition

julia> makeeven(x) = (x ÷ 2) * 2

julia> pos = map(makeeven, Data.Integers{Int8}())

julia> all(iseven, example(pos, 10_000))
true

Supposition.Data.SampledFrom — Type

SampledFrom(collection) <: Possibility{eltype(collection)}

A Possibility for sampling uniformly from collection.

collection, as well as its eachindex, is assumed to be indexable.

Mutable Data

The source objects from the collection given to this SampledFrom is not copied when produce is called. Be careful with mutable data!

julia> using Supposition

julia> sampler = Data.SampledFrom([1, 1, 1, 2])

julia> example(sampler, 4)
4-element Vector{Int64}:
 1
 1
 2
 1

Supposition.Data.Satisfying — Type

Satisfying(source::Possibility, pred) <: Possibility

A Possibility representing values from source fulfilling pred.

Equivalent to calling filter(f, source).

julia> using Supposition

julia> pos = filter(iseven, Data.Integers{Int8}())

julia> all(iseven, example(pos, 10_000))
true

Supposition.Data.Text — Type

Text(alphabet::Possibility{Char}; min_len=0, max_len=10_000) <: Possibility{String}

A Possibility for generating text containing characters of a given alphabet.

julia> using Supposition

julia> text = Data.Text(Data.AsciiCharacters(); max_len=15)

julia> example(text, 5)
5-element Vector{String}:
 "U\x127lxf"
 "hm\x172SJ-("
 "h`\x03\0\x01[[il"
 "\x0ep4"
 "9+Hk3 ii\x1eT"

Supposition.Data.Vectors — Type

Vectors(elements::Possibility{T}; min_size=0, max_size=10_000) <: Possibility{Vector{T}}

A Possibility representing drawing vectors with length l in min_size <= l <= max_size, holding elements of type T.

min_size and max_size must be positive numbers, with min_size <= max_size.

julia> using Supposition

julia> vs = Data.Vectors(Data.Floats{Float16}(); max_size=5)

julia> example(vs, 3)
3-element Vector{Vector{Float16}}:
 [9.64e-5, 9.03e3, 0.04172, -0.0003352]
 [9.793e-5, -2.893, 62.62, 0.0001961]
 [-0.007023, NaN, 3.805, 0.1943]

Supposition reference

Supposition.@check

Supposition.@check

Supposition.@composed

Supposition.@composed

Supposition.adjust

Supposition.assume!

Supposition.choice!

Supposition.consider

Supposition.err_choices

Supposition.example

Supposition.example

Supposition.find_user_stack_depth

Supposition.for_choices

Supposition.forced_choice!

Supposition.generate!

Supposition.reject

Supposition.reject!

Supposition.run

Supposition.should_keep_generating

Supposition.shrink_redistribute

Supposition.shrink_reduce

Supposition.shrink_remove

Supposition.shrink_sort

Supposition.shrink_swap

Supposition.shrink_zeros

Supposition.target!

Supposition.target!

Supposition.test_function

Supposition.weighted!

Supposition.windows

Supposition.BiasedNumbers

Supposition.CheckConfig

Supposition.DirectoryDB

Supposition.Error

Supposition.ExampleDB

Supposition.Fail

Supposition.NoRecordDB

Supposition.Pass

Supposition.Result

Supposition.SuppositionReport

Supposition.TestCase

Supposition.TestState

Supposition.@check — Macro

@check [key=val]... function...

The main way to declare & run a property based test. Called like so:

julia> using Supposition, Supposition.Data

julia> Supposition.@check [options...] function foo(a = Data.Text(Data.Characters(); max_len=10))
          length(a) > 8
       end

Supported options, passed as key=value:

rng::Random.AbstractRNG: Pass an RNG to use. Defaults to Random.Xoshiro(rand(Random.RandomDevice(), UInt)).

max_examples::Int: The maximum number of generated examples that are passed to the property.

broken::Bool: Mark a property that should pass but doesn't as broken, so that failures are not counted.

record::Bool: Whether the result of the invocation should be recorded with any parent testsets.

The arguments to the given function are expected to be generator strategies. The names they are bound to are the names the generated object will have in the test. These arguments will be shown should the property fail!

Extended help

Reusing existing properties

If you already have a predicate defined, you can also use the calling syntax in @check. Here, the generator is passed purely positionally to the given function; no argument name is necessary.

julia> using Supposition, Supposition.Data

julia> isuint8(x) = x isa UInt8

julia> intgen = Data.Integers{UInt8}()

julia> Supposition.@check isuint8(intgen)

Passing a custom RNG

It is possible to optionally give a custom RNG object that will be used for random data generation. If none is given, Xoshiro(rand(Random.RandomDevice(), UInt)) is used instead.

julia> using Supposition, Supposition.Data, Random

# use a custom Xoshiro instance
julia> Supposition.@check rng=Xoshiro(1234) function foo(a = Data.Text(Data.Characters(); max_len=10))
          length(a) > 8
       end

Hardware RNG

Be aware that you cannot pass a hardware RNG to @check directly. If you want to randomize based on hardware entropy, seed a copyable RNG like Xoshiro from your hardware RNG and pass that to @check instead. The RNG needs to be copyable for reproducibility.

Supposition.@composed — Macro

@composed

A way to compose multiple Possibility into one, by applying a function.

The return type is inferred as a best-effort!

Used like so:

julia> using Supposition, Supposition.Data

julia> text = Data.Text(Data.AsciiCharacters(); max_len=10)

julia> gen = Supposition.@composed function foo(a = text, num=Data.Integers(0, 10))
              lpad(num, 2) * ": " * a
       end

julia> example(gen)
" 8:  giR2YL\rl"

Supposition.adjust — Method

adjust(ts::TestState, attempt)

Adjust ts by testing for the choices given by attempt.

Returns whether attempt gave a better score than the best recorded score in ts.

Used exclusively for targeting.

Supposition.assume! — Method

assume!(precondition::Bool)

If this precondition is not met, abort the test and mark the currently running testcase as invalid.

Supposition.choice! — Method

choice!(tc::TestCase, n)

Force a number of choices to occur, taking from the existing prefix first. If the prefix is exhausted, draw from [zero(n), n] instead.

Supposition.consider — Method

consider(ts::TestState, choices) -> Bool

Returns whether the given choices are a conceivable example for the testcase given by ts.

Supposition.err_choices — Method

err_choices

Return the choices that led to the recorded error, if any. If none, return Nothing.

Supposition.example — Method

example(gen::Possibility)

Generate an example for the given Possibility.

julia> using Supposition, Supposition.Data

julia> example(Data.Integers(0, 10))
7

Supposition.example — Method

example(gen::Possibility, n::Integer)

Generate n examples for the given Possibility.

julia> using Supposition, Supposition.Data

julia> is = Data.Integers(0, 10);

julia> example(is, 10)
10-element Vector{Int64}:
  9
  1
  4
  4
  7
  4
  6
 10
  1
  8

Supposition.find_user_stack_depth — Method

find_user_stack_depth

Return a heuristic guess for how many functions deep in user code an error was thrown. Falls back to the full length of the stacktrace.

Supposition.for_choices — Function

for_choices(prefix, rng=Random.default_rng())

Create a TestCase for a given set of known choices.

Supposition.forced_choice! — Method

forced_choice(tc::TestCase, n::UInt64)

Insert a definite choice in the choice sequence.

Note that all integrity checks happen here!

Supposition.generate! — Method

generate(ts::TestState)

Try to generate an example that falsifies the property given to ts.

Supposition.reject! — Method

reject!()

Reject the current testcase as invalid, meaning the generated example should not be considered as producing a valid counterexample.

Supposition.reject — Method

reject(::TestCase)

Mark this test case as invalid.

Supposition.run — Method

run(ts::TestState)

Run the checking algorithm on ts, generating values until we should stop, targeting the score we want to target on and finally shrinking the result.

Supposition.should_keep_generating — Method

should_keep_generating(ts::TestState)

Whether ts should keep generating new test cases, or whether ts is finished.

true returned here means that the given property is not trivial, there is no result yet and we have room for more examples.

Supposition.shrink_redistribute — Method

shrink_redistribute(ts::TestState, attempt::Vector{UInt64}, k::UInt)

Try to shrink attempt by redistributing value between two elements length k apart.

Supposition.shrink_reduce — Method

shrink_reduce(::TestState, attempt::Vector{UInt64})

Try to shrink attempt by making the elements smaller.

Supposition.shrink_remove — Method

shrink_remove(ts::TestState, attempt::Vector{UInt64}, k::UInt)

Try to shrink attempt by removing k elements at a time

Supposition.shrink_sort — Method

shrink_sort(::TestState, attempt::Vector{UInt64}, k::UInt)

Try to shrink attempt by sorting k contiguous elements at a time.

Supposition.shrink_swap — Method

shrink_swap(::TestState, attempt::Vector{UInt64}, k::UInt)

Try to shrink attempt by swapping two elements length k apart.

Supposition.shrink_zeros — Method

shrink_zeros(::TestSTate, attempt::Vector{UInt64}, k::UInt)

Try to shrink attempt by setting k elements at a time to zero.

Supposition.target! — Method

target!(score::Float64)

Update the currently running testcase to track the given score as its target.

score must be convertible to a Float64.

Supposition.target! — Method

target!(ts::TestState)

If ts has a target to go towards set, this will try to climb towards that target by adjusting the choice sequence until ts shouldn't generate anymore.

If ts is currently tracking an error it encountered, it will try to minimize the stacktrace there instead.

Supposition.test_function — Method

test_function(ts::TestState, tc::TestCase)

Test the function given to ts on the test case tc.

Returns a NTuple{Bool, 2} indicating whether tc is interesting and whether it is "better" than the previously best recorded example in ts.

Supposition.weighted! — Method

weighted(tc::TestCase, p::Float64)

Return true with probability p, false otherwise.

Supposition.windows — Method

windows(array, a, b)

Split array into three windows, with split points at a and b. The split points belong to the middle window.

Supposition.BiasedNumbers — Type

BiasedNumbers(weights::Vector{Float64})

Sample the numbers from 1:length(weights), each with a weight of weights[i].

The weights may be any positive number.

Supposition.CheckConfig — Type

CheckConfig

A struct holding the initial configuration for an invocation of @check.

Fields:

rng: The initial RNG object given to @check

max_examples: The maximum number of examples allowed to be drawn with this config

record: Whether the result should be recorded in the parent testset, if there is one

Supposition.DirectoryDB — Type

DirectoryDB <: ExampleDB

An ExampleDB that records examples as files in a directory.

Supposition.Error — Type

Error

A result indicating that an error was encountered while generating or shrinking.

Supposition.Fail — Type

Fail

A result indicating that a counterexample was found.

Supposition.NoRecordDB — Type

NoRecordDB <: ExambleDB

An ExampleDB that doesn't record anything, and won't retrieve anything.

Supposition.Pass — Type

Pass

A result indicating that no counterexample was found.

Supposition.Result — Type

Result

An abstract type representing the ultimate result a TestState ended up at.

Supposition.SuppositionReport — Type

SuppositionReport <: AbstractTestSet

An AbstractTestSet, for recording the final result of @check in the context of @testset

Supposition.TestCase — Type

TestCase

A struct representing a single (ongoing) test case.

prefix: A fixed set of choices that must be made first.

rng: The RNG this testcase ultimately uses to draw from. This is used to seed the task-local RNG object before generating begins.

max_size: The maximum number of choices this TestCase is allowed to make.

choices: The binary choices made so far.

targeting_score: The score this TestCase attempts to maximize.

Supposition.TestState — Type

TestState

config: The configuration this TestState is running with

is_interesting: The user given property to investigate

rng: The currently used RNG

valid_test_cases: The count of (so far) valid encountered testcases

result: The choice sequence leading to a non-throwing counterexample

best_scoring: The best scoring result that was encountered during targeting

target_err: The error this test has previously encountered and the smallest choice sequence leading to it

test_is_trivial: Whether is_interesting is trivial, i.e. led to no choices being required